+++ src/types/ink.types.ts
@@ -31,6 +31,8 @@ export type ToolbarOption =
   | 'link'
   | 'image'
   | 'table'
+  | 'signature'
+  | 'findReplace'
   | 'undo'
   | 'redo'
   | 'trackChanges'
@@ -83,6 +85,8 @@ export interface InkEditorProps extends Omit<HTMLAttributes<HTMLDivElement>, 'on
   pasteMode?: InkPasteMode;
   onImageUpload?: InkImageUploadHandler;
   wysiwyg?: boolean;
+  keepInMemory?: boolean;
+  memoryKey?: string;
 }
 
 export interface ToolbarButtonProps {